refactor: moved two functions from piece file to utils file
