+++ src/utils.hpp
@@ -5,5 +5,7 @@
 
 
 Color get_opposite_color(Color color);
+char get_char_representation(PieceType piece_type);
+PieceType get_piece_type(char char_representation);
 
 #endif
